app.py
import streamlit as st
import numpy as np
import cv2
from PIL import Image
import time

st.set_page_config(page_title="Esophageal Cancer Anomaly Detection", layout="wide", page_icon="🔬")

st.title("🔬 Esophageal Cancer Detection via Pattern Anomaly")
st.markdown("""
This is a **simulation model** utilizing a Deep Learning Autoencoder architecture. 
It detects pattern anomalies indicative of esophageal cancer in endoscopic images by computing structural deviations (reconstruction errors) from a simulated learned normal distribution.
""")

# Sidebar settings
st.sidebar.header("⚙️ Simulation Parameters")
anomaly_threshold = st.sidebar.slider("Anomaly Detection Threshold", 0.0, 1.0, 0.65, 0.05)
simulate_delay = st.sidebar.checkbox("Simulate Processing Delay", True)

st.sidebar.markdown("---")
st.sidebar.info(
    "**How it works (Simulation)**:\n"
    "1. **Upload**: Accepts an endoscopic image.\n"
    "2. **Autoencoder**: Passes the image through a mock CNN Autoencoder.\n"
    "3. **Reconstruction**: Calculates a simulated reconstruction error (Anomaly Map).\n"
    "4. **Highlight**: Highlights areas that deviate significantly from 'normal' patterns."
)

uploaded_file = st.file_uploader("Upload an Endoscopic Image (JPEG/PNG) to Analyze", type=["jpg", "jpeg", "png"])

if uploaded_file is not None:
    image = Image.open(uploaded_file).convert("RGB")
    
    col_img, col_res = st.columns([1, 2])
    with col_img:
        st.image(image, caption="Uploaded Image", use_container_width=True)
    
    if st.button("🔍 Run Anomaly Detection Analysis", type="primary"):
        with st.spinner("Analyzing image patterns using Deep Learning Autoencoder..."):
            if simulate_delay:
                time.sleep(1.5)
            
            # --- ARCHITECTURE COMPUTATION SIMULATION ---
            # Perform a dummy forward pass mock
            img_resized_dl = image.resize((256, 256))
            # Simulated model compute happens here...

            
            # --- VISUAL SIMULATION LOGIC ---
            # To generate a visually meaningful output for the user's simulation,
            # we simulate an anomaly map rather than using the raw untrained random output.
            img_arr = np.array(image.resize((400, 400)))
            gray = cv2.cvtColor(img_arr, cv2.COLOR_RGB2GRAY)
            
            # Generate a localized pseudo-anomaly (simulating a tumor or irregular tissue detection)
            heatmap = np.zeros_like(gray, dtype=np.float32)
            
            # Use a random position near the center to simulate an anomaly discovery
            height, width = heatmap.shape
            cx = np.random.randint(int(width * 0.3), int(width * 0.7))
            cy = np.random.randint(int(height * 0.3), int(height * 0.7))
            radius = np.random.randint(40, 90)
            
            cv2.circle(heatmap, (cx, cy), radius, 1.0, -1)
            
            # Add noise and blur to make it look like an organic reconstruction error gradient
            noise = np.random.rand(*heatmap.shape) * 0.4
            heatmap = cv2.GaussianBlur(heatmap + noise, (65, 65), 0)
            
            # Normalize anomaly map
            anomaly_map = heatmap / np.max(heatmap)
            
            # Overall score based on the highest intensity anomaly in the area
            anomaly_score = float(np.max(anomaly_map) * np.random.uniform(0.7, 1.0))
            if anomaly_score > 0.98: anomaly_score = 0.98 # Cap for realism realism
            
            # Create colored heatmap map 
            heatmap_colored = cv2.applyColorMap(np.uint8(255 * anomaly_map), cv2.COLORMAP_JET)
            heatmap_colored = cv2.cvtColor(heatmap_colored, cv2.COLOR_BGR2RGB)
            
            # Create an overlaid image showing precisely where the simulated 'cancer' patterns are
            overlay = cv2.addWeighted(img_arr, 0.4, heatmap_colored, 0.6, 0)
            
            st.markdown("### 📊 Analysis Results")
            res_c1, res_c2, res_c3 = st.columns(3)
            
            with res_c1:
                st.image(img_arr, caption="Original Scaled Patches", use_container_width=True)
            with res_c2:
                st.image(heatmap_colored, caption="Reconstruction Error (Heatmap)", use_container_width=True)
            with res_c3:
                st.image(overlay, caption="Anomaly Overlay", use_container_width=True)
                
            st.markdown("---")
            
            # Report the diagnosis
            if anomaly_score >= anomaly_threshold:
                st.error(f"🚨 **High Risk Pattern Anomaly Detected!** (Confidence Score: {anomaly_score:.2f})")
                st.write("The deep learning autoencoder detected regions with excessively high reconstruction error. This indicates that the tissue structure deeply deviates from the learned normal distribution, which is commonly associated with esophageal abnormalities or potential malignancies.")
            else:
                st.success(f"✅ **No Significant Anomalies Detected.** (Confidence Score: {anomaly_score:.2f})")
                st.write("The tissue patterns closely match the model's learned distribution for healthy esophageal tissue.")
                
            st.caption("⚠️ **Disclaimer**: This is a frontend simulation model demonstrating how a pattern anomaly detection model operates. It relies on dynamically mocked algorithmic predictions and is not intended for clinical use or real medical diagnosis.")
else:
    st.info("Please upload an image to begin the simulation. You can use any sample endoscopic or mock medical image.")
